Dropping a db fails consistely across adapters.
Resolves issue: #32914. When dropping a DB on sqlite3 and on postgresql
if the database (say 'development' and 'test') did not exist the error messages
where inconsistent. On sqlite it would say something like:
```
$ rake db:drop
Dropped database 'db/development.sqlite3'
Database 'db/test.sqlite3' does not exist
```

This was also occuring for postgresql as well. This patch makes the
required fixes to make this experience consistent across adapdters.
After this patch dropping a non existent database should look like:

```
$ rake db:drop
Database 'db/test.sqlite3' does not exist
Database 'db/development.sqlite3' does not exist
```

Across all three: sqlite3, postgres, and mysql adapters.

+++ activerecord/lib/active_record/tasks/postgresql_database_tasks.rb
@@ -29,7 +29,15 @@ module ActiveRecord
         end
       end
 
+      def database_exists?
+        !!connection
+      rescue ActiveRecord::NoDatabaseError
+        false
+      end
+
       def drop
+        raise ActiveRecord::NoDatabaseError unless database_exists?
+
         establish_master_connection
         connection.drop_database configuration["database"]
       end
